+++ src/core/mangler_context.ts
@@ -1,8 +1,13 @@
 import { CSSMangleReserved } from "../types";
 
+export interface ManglerContextOptions {
+    reversed: CSSMangleReserved;
+    canUndeclared: boolean,
+}
+
 export class ManglerContext<T> {
     constructor(
-        public options: {reversed: CSSMangleReserved},
+        public options: ManglerContextOptions,
         public parent: T
     ) {}
 }