+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 
 # 1.0.0-alpha7
 - Fixed to transform variables using even upper-case alphabets.
+- Added undeclared into options property, this is whether an undeclared identifier in this webpack plugin would still be considered for minification.